+++ android/anc/src/test/java/org/smartregister/fhircore/anc/shadow/AncApplicationShadow.kt
@@ -119,4 +119,7 @@ open class AncApplicationShadow : ShadowApplication() {
     field.isAccessible = true
     return field.get(search) as MutableList<T>
   }
+  @Implementation fun schedulePeriodicSync() {}
+
+  @Implementation fun initializeWorkerContextProvider() {}
 }