+++ android/eir/src/test/java/org/smartregister/fhircore/eir/shadow/EirApplicationShadow.kt
@@ -119,4 +119,7 @@ class EirApplicationShadow : ShadowApplication() {
     field.isAccessible = true
     return field.get(search) as MutableList<T>
   }
+  @Implementation fun schedulePeriodicSync() {}
+
+  @Implementation fun initializeWorkerContextProvider() {}
 }