+++ android/eir/src/test/java/org/smartregister/fhircore/eir/ui/patient/register/PatientRegisterFragmentTest.kt
@@ -33,11 +33,12 @@ import org.smartregister.fhircore.eir.data.model.VaccineStatus
 import org.smartregister.fhircore.eir.robolectric.RobolectricTest
 import org.smartregister.fhircore.eir.shadow.EirApplicationShadow
 import org.smartregister.fhircore.eir.shadow.FakeKeyStore
+import org.smartregister.fhircore.eir.shadow.ShadowNpmPackageProvider
 import org.smartregister.fhircore.eir.ui.patient.details.PatientDetailsActivity
 import org.smartregister.fhircore.eir.ui.vaccine.RecordVaccineActivity
 import org.smartregister.fhircore.engine.ui.register.model.RegisterFilterType
 
-@Config(shadows = [EirApplicationShadow::class])
+@Config(shadows = [EirApplicationShadow::class, ShadowNpmPackageProvider::class])
 class PatientRegisterFragmentTest : RobolectricTest() {
 
   private lateinit var registerFragment: PatientRegisterFragment