+++ CHANGELOG.md
@@ -12,10 +12,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added measure reporting to ANC application
 - Added class for Measure report evaluation which will be used in ANC application
 - ANC | Added Condition resource to sync params list
+- Moved Token to secure storage from AccountManager
+- Expose [custom user attribute](https://www.keycloak.org/docs/latest/server_admin/index.html#_user-attributes) `questionnaire_publisher` available in SharedPreferences with key `USER_QUESTIONNAIRE_PUBLISHER` (#607)
+- Quest | Filter [Questionnaires](http://hl7.org/fhir/questionnaire.html) by [publisher](http://hl7.org/fhir/questionnaire-definitions.html#Questionnaire.publisher) using user attribute as per above. (#571)
+- Quest | Patient List, Load Config from server
+- Quest | Added Patient Profile View
+- Quest | Patient Registration Questionnaire
 
 ### Fixed
 
 - Added dependecies that were missing and causing CQL Evaluation to fail
+- Out of memory issue on few tests
+- Authentication toekn expiry issue
+- Fhir Resource Converter issue after resource update
+- Inteceptor handling for missing account
 
 ### Changed
 
@@ -39,7 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refactor COVAX Immunization extraction to use StructureMap-based extractio
 
-## [0.0.8 EIR, 0.0.1 - ANC] - 2021-09-27
+## [0.0.8 EIR, 0.0.1 - ANC] - 2021-09-11
 
 ### Added